+++ packages/web-app/src/utils/documentTypeHelpers.js
@@ -188,4 +188,3 @@ export const documentTypeHelpers = {
   isTopographicDrawing,
   isUnknown
 };
-
