+++ packages/web-app/src/utils/formatNotification.js
