+++ packages/web-app/src/utils/strings.js
